refactor(server): asr,common,llm,tts,util,vad file struct
