+++ apps/server/api/src/common/mod.rs
