+++ apps/server/api/src/llm/chat.rs
