+++ apps/server/api/src/llm/models/mod.rs
