+++ apps/server/api/src/llm/models/qwen3/mod.rs
@@ -7,7 +7,7 @@ extern crate accelerate_src;
 pub mod quantized;
 
 use super::token_output_stream::TokenOutputStream;
-use crate::ws::{
+use crate::{
     common::{ModelError, device, format_size},
     llm::{Model, models::token_converter::TokenConverter},
 };
@@ -385,7 +385,7 @@ mod tests {
     use tracing::info;
     use tracing_test::traced_test;
 
-    use crate::ws::llm::models::qwen3::convert_request_to_prompt;
+    use crate::llm::models::qwen3::convert_request_to_prompt;
 
     #[tokio::test]
     #[traced_test]
