+++ apps/server/api/src/llm/models/token_output_stream.rs
