+++ apps/server/api/src/util/llm.rs
