+++ apps/server/api/src/ws/message_converter.rs
@@ -1,4 +1,4 @@
-use crate::ws::frame::Frame;
+use super::frame::Frame;
 use axum::extract::ws::Message;
 use serde_json::Value;
 use service::chobits::message::{
