+++ apps/server/api/src/ws/mod.rs
@@ -1,27 +1,19 @@
-pub mod asr;
-pub mod common;
 pub mod frame;
-pub mod llm;
 pub mod message_converter;
 pub mod session;
-pub mod tts;
-pub mod util;
-pub mod vad;
 
 use crate::{
-    AppState, config,
+    AppState,
+    asr::asr_cache::AsrCache,
+    config,
+    llm::LlmFactory,
     mcp::{
         client::server::ServerMcpClient,
         mcp_host::{McpHost, UnionMcpHost},
     },
-    ws::{
-        asr::asr_cache::AsrCache,
-        llm::LlmFactory,
-        message_converter::convert_to_frame,
-        session::{SessionBuilder, SessionConfig, listener::DefaultListener},
-        vad::vad_cache::VadCache,
-    },
+    vad::vad_cache::VadCache,
 };
+
 use axum::{
     RequestPartsExt, debug_handler,
     extract::{ConnectInfo, FromRequestParts, Path, WebSocketUpgrade, ws::Message},
@@ -31,10 +23,12 @@ use axum::{
 use axum_extra::{TypedHeader, headers};
 use framework::id::gen_id;
 use futures_util::{Sink, SinkExt, Stream, StreamExt};
+use message_converter::convert_to_frame;
 use rmcp::transport::{
     StreamableHttpClientTransport, streamable_http_client::StreamableHttpClientTransportConfig,
 };
 use serde::Serialize;
+use session::{SessionBuilder, SessionConfig, listener::DefaultListener};
 use std::{collections::HashMap, net::SocketAddr, sync::Arc};
 use tokio::sync::Mutex;
 use tracing::{error, info};
