+++ apps/server/api/src/ws/session/listener.rs
@@ -1,7 +1,4 @@
-use crate::{
-    config,
-    ws::{asr::Asr, common::ModelError, vad::Vad},
-};
+use crate::{asr::Asr, common::ModelError, config, vad::Vad};
 use async_trait::async_trait;
 use chrono::Local;
 use std::sync::Arc;
