+++ apps/server/api/tests/audio_test.rs
@@ -1,4 +1,4 @@
-use api::ws::util::audio::pcm_decode;
+use api::util::audio::pcm_decode;
 use std::cmp;
 use std::path::PathBuf;
 use tracing::info;
