+++ apps/server/api/tests/session_test.rs
@@ -9,24 +9,22 @@ use std::{
 
 use api::{
     AppState, config,
+    llm::LlmFactory,
     mcp::{
         client::server::ServerMcpClient,
         mcp_host::{McpHost, UnionMcpHost},
     },
     setup_mcp,
-    ws::{
-        llm::LlmFactory,
-        session::{SessionBuilder, SessionConfig},
-        tts::TtsFactory,
-        util::audio::pcm_decode,
-    },
+    tts::TtsFactory,
+    util::audio::pcm_decode,
+    ws::session::{SessionBuilder, SessionConfig},
 };
 
-use api::ws::{
+use api::{
     asr::asr_cache::AsrCache,
-    frame::{Frame, FrameResult},
-    session::{Session, listener::DefaultListener},
     vad::vad_cache::VadCache,
+    ws::frame::{Frame, FrameResult},
+    ws::session::{Session, listener::DefaultListener},
 };
 use framework::id::gen_id;
 use rmcp::{
